Extract AbstracttrackingListener for Blameable, IpTraceble and Timestampable listeners
